requirements.txt
# ── Core ──────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9      # required for Form() and File() uploads
python-dotenv>=1.0.1

# ── Database ──────────────────────────────────
motor>=3.4.0                  # async MongoDB driver
pymongo>=4.7.0

# ── Auth ──────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# ── AI / LLM ──────────────────────────────────
openai>=1.30.0                # AsyncOpenAI client (Featherless-compatible)

# ── TurboLearn: transcription ─────────────────
yt-dlp>=2024.5.1              # YouTube audio download
openai-whisper>=20231117      # local Whisper fallback
pydub>=0.25.1                 # audio utilities

# ── TurboLearn: PDF ───────────────────────────
PyMuPDF>=1.24.0               # fitz — PDF text extraction

# ── ML ────────────────────────────────────────
scikit-learn>=1.4.0
numpy>=1.26.0

# ── Dev / test ────────────────────────────────
psutil>=5.9.0
pytest>=8.0.0
pytest-anyio>=0.0.0
anyio>=4.0.0
httpx>=0.27.0
